+++ src/migrations/M190904193339Shopify.php
@@ -2,15 +2,15 @@
 
 declare(strict_types=1);
 
-namespace Hirtz\Shopify\migrations;
+namespace Hirtz\Shopify\Migrations;
 
 use Hirtz\Shopify\models\Product;
 use Hirtz\Shopify\models\ProductImage;
 use Hirtz\Shopify\models\ProductVariant;
 use Hirtz\Shopify\models\Webhook;
 use Hirtz\Shopify\modules\ModuleTrait;
-use Hirtz\Skeleton\db\traits\MigrationTrait;
-use Hirtz\Skeleton\models\User;
+use Hirtz\Skeleton\Db\Traits\MigrationTrait;
+use Hirtz\Skeleton\Models\User;
 use Yii;
 use yii\db\Migration;