+++ src/migrations/M240328135307ProductVariant.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Hirtz\Shopify\migrations;
+namespace Hirtz\Shopify\Migrations;
 
 use Hirtz\Shopify\models\ProductVariant;
-use Hirtz\Skeleton\db\traits\MigrationTrait;
+use Hirtz\Skeleton\Db\Traits\MigrationTrait;
 use yii\db\Migration;
 
 /**