+++ tests/fixtures/UserFixture.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Hirtz\Shopify\tests\fixtures;
 
-use Hirtz\Skeleton\models\User;
+use Hirtz\Skeleton\Models\User;
 use yii\test\ActiveFixture;
 
 class UserFixture extends ActiveFixture